perf(combobox): prevent initial list render and update tests to prove that reduces render time
