+++ packages/combobox/test/benchmark/light-dom-test.ts
@@ -16,14 +16,19 @@ import { html } from '@spectrum-web-components/base';
 import { measureFixtureCreation } from '../../../../test/benchmark/helpers.js';
 import { countryList } from '../index.js';
 
-measureFixtureCreation(html`
-    <sp-combobox>
-        ${countryList.map(
-            (option, index) => html`
-                <sp-menu-item id=${index} value=${option}>
-                    ${option}
-                </sp-menu-item>
-            `
-        )}
-    </sp-combobox>
-`);
+measureFixtureCreation(
+    html`
+        <sp-combobox>
+            ${countryList.map(
+                (option, index) => html`
+                    <sp-menu-item id=${index} value=${option}>
+                        ${option}
+                    </sp-menu-item>
+                `
+            )}
+        </sp-combobox>
+    `,
+    {
+        numRenders: 10,
+    }
+);
