+++ packages/combobox/test/index.ts
@@ -10,13 +10,6 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { expect, fixture } from '@open-wc/testing';
-import { html } from '@open-wc/testing';
-import { ComboboxOption } from '@spectrum-web-components/combobox';
-import '@spectrum-web-components/combobox/sp-combobox.js';
-import { MenuItem } from '@spectrum-web-components/menu';
-import { fruits } from '../stories/index.js';
-
 export const countryList = [
     'Afghanistan',
     'Albania',
@@ -273,47 +266,3 @@ export const benchmarkOptions = countryList.map((value, index) => ({
     value: index.toString(),
     itemText: value,
 }));
-
-export type TestableCombobox = HTMLElement & {
-    activeDescendant: ComboboxOption;
-    autocomplete: 'none' | 'list';
-    availableOptions: ComboboxOption[];
-    focused: boolean;
-    focusElement: HTMLInputElement;
-    open: boolean;
-    optionEls: MenuItem[];
-    options: ComboboxOption[];
-    shadowRoot: ShadowRoot;
-    value: string;
-};
-
-export const comboboxFixture = async (): Promise<TestableCombobox> => {
-    const el = await fixture<TestableCombobox>(
-        html`
-            <sp-combobox
-                .autocomplete=${'list'}
-                label="Combobox"
-                .options=${fruits}
-            >
-                Combobox
-            </sp-combobox>
-        `
-    );
-
-    return el;
-};
-
-export const testActiveElement = (
-    el: TestableCombobox,
-    testId: string
-): void => {
-    expect(el.activeDescendant?.value).to.equal(testId);
-    const activeElement = el.shadowRoot.querySelector(
-        `#${el.activeDescendant.value}`
-    ) as HTMLElement;
-    expect(activeElement.getAttribute('aria-selected')).to.equal('true');
-};
-
-export const isWebKit =
-    /AppleWebKit/.test(window.navigator.userAgent) &&
-    !/Chrome/.test(window.navigator.userAgent);
